Move all validation code into one file
- Collect all validation logic in one place for further refactoring.
- Add chain.validate_tx() needed to implement new micro blocks.

This patch is pure refactoring and doesn't cause any functional
changes.
