+++ blockchain/src/block.rs
@@ -21,18 +21,12 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-use crate::error::*;
 use crate::merkle::*;
 use crate::output::*;
 use crate::view_changes::ViewChangeProof;
 use bitvector::BitVector;
-use failure::Error;
 use std::time::SystemTime;
-use stegos_crypto::bulletproofs::{fee_a, validate_range_proof};
-use stegos_crypto::curve1174::cpt::Pt;
-use stegos_crypto::curve1174::ecpt::ECp;
 use stegos_crypto::curve1174::fields::Fr;
-use stegos_crypto::curve1174::G;
 use stegos_crypto::hash::{Hash, Hashable, Hasher};
 use stegos_crypto::pbc::secure;
 use stegos_crypto::pbc::secure::VRF;
@@ -321,69 +315,6 @@ impl MicroBlock {
 
         block
     }
-
-    ///
-    /// Validate the block monetary balance.
-    ///
-    /// This function is a lightweight version of Blockchain.validate_micro_block().
-    /// The only monetary balance is validated. For test purposes only.
-    ///
-    /// # Arguments
-    ///
-    /// * - `inputs` - UTXOs referred by self.body.inputs, in the same order as in self.body.inputs.
-    ///
-    pub fn validate_balance(&self, inputs: &[Output]) -> Result<(), Error> {
-        //
-        // Calculate the pedersen commitment difference in order to check the monetary balance:
-        //
-        //     pedersen_commitment_diff = monetary_adjustment + \sum C_i - \sum C_o
-        //
-
-        let mut pedersen_commitment_diff: ECp = fee_a(self.header.monetary_adjustment);
-
-        // +\sum{C_i} for i in txins
-        for (txin_hash, txin) in self.body.inputs.iter().zip(inputs) {
-            assert_eq!(Hash::digest(txin), *txin_hash);
-            match txin {
-                Output::PaymentOutput(o) => {
-                    pedersen_commitment_diff += Pt::decompress(o.proof.vcmt)?;
-                }
-                Output::StakeOutput(o) => {
-                    pedersen_commitment_diff += fee_a(o.amount);
-                }
-            };
-        }
-
-        // -\sum{C_o} for o in txouts
-        for (txout, _) in self.body.outputs.leafs() {
-            let output_hash = Hash::digest(&*txout);
-            match **txout {
-                Output::PaymentOutput(ref o) => {
-                    // Check bulletproofs of created outputs
-                    if !validate_range_proof(&o.proof) {
-                        return Err(OutputError::InvalidBulletProof(output_hash).into());
-                    }
-                    pedersen_commitment_diff -= Pt::decompress(o.proof.vcmt)?;
-                }
-                Output::StakeOutput(ref o) => {
-                    if o.amount <= 0 {
-                        return Err(OutputError::InvalidStake(output_hash).into());
-                    }
-                    pedersen_commitment_diff -= fee_a(o.amount);
-                }
-            };
-        }
-
-        // Check the monetary balance
-        if pedersen_commitment_diff != self.header.gamma * (*G) {
-            let block_hash = Hash::digest(&self);
-            return Err(
-                BlockError::InvalidBlockBalance(self.header.base.height, block_hash).into(),
-            );
-        }
-
-        Ok(())
-    }
 }
 
 impl Hashable for MicroBlock {
@@ -415,343 +346,4 @@ impl Hashable for Block {
             Block::MicroBlock(micro_block) => micro_block.hash(state),
         }
     }
-}
-
-#[cfg(test)]
-pub mod tests {
-    use super::*;
-    use crate::mix;
-    use std::time::SystemTime;
-    use stegos_crypto::curve1174::cpt::make_random_keys;
-    use stegos_crypto::pbc::secure::make_random_keys as make_secure_random_keys;
-
-    #[test]
-    fn create_validate_micro_block() {
-        let (skey0, _pkey0) = make_random_keys();
-        let (skey1, pkey1) = make_random_keys();
-        let (_skey2, pkey2) = make_random_keys();
-        let (pbc_skey, pbc_pkey) = make_secure_random_keys();
-
-        let version: u64 = 1;
-        let height: u64 = 0;
-        let timestamp = SystemTime::now();
-        let view_change = 0;
-        let amount: i64 = 1_000_000;
-        let previous = Hash::digest("test");
-        let seed = mix(Hash::zero(), view_change);
-        let random = secure::make_VRF(&pbc_skey, &seed);
-
-        //
-        // Valid block with transaction from 1 to 2
-        //
-        {
-            let (output0, gamma0) = Output::new_payment(timestamp, &skey0, &pkey1, amount).unwrap();
-            let base =
-                BaseBlockHeader::new(version, previous, height, view_change, timestamp, random);
-            let inputs1 = [Hash::digest(&output0)];
-            let (output1, gamma1) = Output::new_payment(timestamp, &skey1, &pkey2, amount).unwrap();
-            let outputs1 = [output1];
-            let gamma = gamma0 - gamma1;
-            let block = MicroBlock::new(
-                base, gamma, 0, &inputs1, &outputs1, None, pbc_pkey, &pbc_skey,
-            );
-            block.validate_balance(&[output0]).expect("block is valid");
-        }
-
-        //
-        // Block with invalid monetary balance
-        //
-        {
-            let (output0, gamma0) = Output::new_payment(timestamp, &skey0, &pkey1, amount).unwrap();
-            let base =
-                BaseBlockHeader::new(version, previous, height, view_change, timestamp, random);
-            let inputs1 = [Hash::digest(&output0)];
-            let (output1, gamma1) =
-                Output::new_payment(timestamp, &skey1, &pkey2, amount - 1).unwrap();
-            let outputs1 = [output1];
-            let gamma = gamma0 - gamma1;
-            let block = MicroBlock::new(
-                base, gamma, 0, &inputs1, &outputs1, None, pbc_pkey, &pbc_skey,
-            );
-            match block.validate_balance(&[output0]) {
-                Err(e) => match e.downcast::<BlockError>().unwrap() {
-                    BlockError::InvalidBlockBalance(_height, _hash) => {}
-                    _ => panic!(),
-                },
-                _ => panic!(),
-            }
-        }
-    }
-
-    #[test]
-    fn validate_pruned_micro_block() {
-        let (skey, pkey) = make_random_keys();
-        let (pbc_skey, pbc_pkey) = make_secure_random_keys();
-
-        let version: u64 = 1;
-        let height: u64 = 0;
-        let timestamp = SystemTime::now();
-        let view_change = 0;
-        let amount: i64 = 1_000_000;
-        let previous = Hash::digest(&"test".to_string());
-
-        let seed = mix(Hash::zero(), view_change);
-        let random = secure::make_VRF(&pbc_skey, &seed);
-
-        let (input, gamma0) = Output::new_payment(timestamp, &skey, &pkey, amount).unwrap();
-        let base = BaseBlockHeader::new(version, previous, height, view_change, timestamp, random);
-        let input_hashes = [Hash::digest(&input)];
-        let inputs = [input];
-        let (output, gamma1) = Output::new_payment(timestamp, &skey, &pkey, amount).unwrap();
-        let outputs = [output];
-        let gamma = gamma0 - gamma1;
-        let block = MicroBlock::new(
-            base,
-            gamma,
-            0,
-            &input_hashes,
-            &outputs,
-            None,
-            pbc_pkey,
-            &pbc_skey,
-        );
-        block.validate_balance(&inputs).expect("block is valid");
-
-        {
-            // Prune an output.
-            let mut block2 = block.clone();
-            let (_output, path) = block2.body.outputs.leafs()[0];
-            block2.body.outputs.prune(&path).expect("output exists");
-            match block2.validate_balance(&inputs) {
-                Err(e) => match e.downcast::<BlockError>().unwrap() {
-                    BlockError::InvalidBlockBalance(_height, _hash) => {}
-                    _ => panic!(),
-                },
-                _ => panic!(),
-            }
-        }
-    }
-
-    #[test]
-    fn create_validate_micro_block_with_escrow() {
-        let (skey0, _pkey0) = make_random_keys();
-        let (skey1, pkey1) = make_random_keys();
-        let (secure_skey1, secure_pkey1) = make_secure_random_keys();
-
-        let version: u64 = 1;
-        let height: u64 = 0;
-        let timestamp = SystemTime::now();
-        let view_change = 0;
-        let amount: i64 = 1_000_000;
-        let previous = Hash::digest(&"test".to_string());
-        let seed = mix(Hash::zero(), view_change);
-        let random = secure::make_VRF(&secure_skey1, &seed);
-
-        //
-        // Escrow as an input.
-        //
-        {
-            let input = Output::new_stake(
-                timestamp,
-                &skey0,
-                &pkey1,
-                &secure_pkey1,
-                &secure_skey1,
-                amount,
-            )
-            .expect("keys are valid");
-            let input_hashes = [Hash::digest(&input)];
-            let inputs = [input];
-            let inputs_gamma = Fr::zero();
-            let (output, outputs_gamma) =
-                Output::new_payment(timestamp, &skey1, &pkey1, amount).expect("keys are valid");
-            let outputs = [output];
-            let gamma = inputs_gamma - outputs_gamma;
-
-            let base =
-                BaseBlockHeader::new(version, previous, height, view_change, timestamp, random);
-            let block = MicroBlock::new(
-                base,
-                gamma,
-                0,
-                &input_hashes[..],
-                &outputs[..],
-                None,
-                secure_pkey1,
-                &secure_skey1,
-            );
-            block.validate_balance(&inputs).expect("block is valid");
-        }
-
-        //
-        // Escrow as an output.
-        //
-        {
-            let (input, inputs_gamma) =
-                Output::new_payment(timestamp, &skey0, &pkey1, amount).expect("keys are valid");
-            let input_hashes = [Hash::digest(&input)];
-            let inputs = [input];
-            let output = Output::new_stake(
-                timestamp,
-                &skey1,
-                &pkey1,
-                &secure_pkey1,
-                &secure_skey1,
-                amount,
-            )
-            .expect("keys are valid");
-            let outputs_gamma = Fr::zero();
-            let outputs = [output];
-            let gamma = inputs_gamma - outputs_gamma;
-
-            let base =
-                BaseBlockHeader::new(version, previous, height, view_change, timestamp, random);
-            let block = MicroBlock::new(
-                base,
-                gamma,
-                0,
-                &input_hashes[..],
-                &outputs[..],
-                None,
-                secure_pkey1,
-                &secure_skey1,
-            );
-            block.validate_balance(&inputs).expect("block is valid");
-        }
-
-        //
-        // Invalid monetary balance.
-        //
-        {
-            let (input, inputs_gamma) =
-                Output::new_payment(timestamp, &skey0, &pkey1, amount).expect("keys are valid");
-            let input_hashes = [Hash::digest(&input)];
-            let inputs = [input];
-            let mut output = StakeOutput::new(
-                timestamp,
-                &skey1,
-                &pkey1,
-                &secure_pkey1,
-                &secure_skey1,
-                amount,
-            )
-            .expect("keys are valid");
-            output.amount = amount - 1;
-            let output = Output::StakeOutput(output);
-            let outputs_gamma = Fr::zero();
-            let outputs = [output];
-            let gamma = inputs_gamma - outputs_gamma;
-
-            let base =
-                BaseBlockHeader::new(version, previous, height, view_change, timestamp, random);
-            let block = MicroBlock::new(
-                base,
-                gamma,
-                0,
-                &input_hashes[..],
-                &outputs[..],
-                None,
-                secure_pkey1,
-                &secure_skey1,
-            );
-            match block.validate_balance(&inputs) {
-                Err(e) => match e.downcast::<BlockError>().unwrap() {
-                    BlockError::InvalidBlockBalance(_height, _hash) => {}
-                    _ => panic!(),
-                },
-                _ => panic!(),
-            };
-        }
-
-        //
-        // Invalid stake.
-        //
-        {
-            let (input, inputs_gamma) =
-                Output::new_payment(timestamp, &skey0, &pkey1, amount).expect("keys are valid");
-            let input_hashes = [Hash::digest(&input)];
-            let inputs = [input];
-            let mut output = StakeOutput::new(
-                timestamp,
-                &skey1,
-                &pkey1,
-                &secure_pkey1,
-                &secure_skey1,
-                amount,
-            )
-            .expect("keys are valid");
-            output.amount = 0;
-            let output = Output::StakeOutput(output);
-            let outputs_gamma = Fr::zero();
-            let outputs = [output];
-            let gamma = inputs_gamma - outputs_gamma;
-
-            let base =
-                BaseBlockHeader::new(version, previous, height, view_change, timestamp, random);
-            let block = MicroBlock::new(
-                base,
-                gamma,
-                0,
-                &input_hashes[..],
-                &outputs[..],
-                None,
-                secure_pkey1,
-                &secure_skey1,
-            );
-            match block.validate_balance(&inputs) {
-                Err(e) => match e.downcast::<OutputError>().unwrap() {
-                    OutputError::InvalidStake(_output_hash) => {}
-                    _ => panic!(),
-                },
-                _ => panic!(),
-            };
-        }
-    }
-
-    fn create_burn_money(input_amount: i64, output_amount: i64) {
-        let (skey, pkey) = make_random_keys();
-        let (secure_skey1, secure_pkey1) = make_secure_random_keys();
-
-        let version: u64 = 1;
-        let height: u64 = 0;
-        let timestamp = SystemTime::now();
-        let view_change = 0;
-        let previous = Hash::digest(&"test".to_string());
-
-        let seed = mix(Hash::zero(), view_change);
-        let random = secure::make_VRF(&secure_skey1, &seed);
-        let monetary_adjustment: i64 = output_amount - input_amount;
-
-        let (input, input_gamma) =
-            Output::new_payment(timestamp, &skey, &pkey, input_amount).unwrap();
-        let base = BaseBlockHeader::new(version, previous, height, view_change, timestamp, random);
-        let input_hashes = [Hash::digest(&input)];
-        let inputs = [input];
-        let (output, output_gamma) =
-            Output::new_payment(timestamp, &skey, &pkey, output_amount).unwrap();
-        let outputs = [output];
-        let gamma = input_gamma - output_gamma;
-        let block = MicroBlock::new(
-            base,
-            gamma,
-            monetary_adjustment,
-            &input_hashes,
-            &outputs,
-            None,
-            secure_pkey1,
-            &secure_skey1,
-        );
-        block.validate_balance(&inputs).expect("block is valid");
-    }
-
-    #[test]
-    fn create_money() {
-        create_burn_money(100, 200);
-    }
-
-    #[test]
-    fn burn_money() {
-        create_burn_money(200, 100);
-    }
-
 }
